+++ src/main/java/com/ats/webapi/repository/StockCalculationRepository.java
@@ -16,7 +16,7 @@ public interface StockCalculationRepository extends JpaRepository<RegularSpecial
 			@Param("toDate") String toDate, @Param("itemId") int itemId);
 
 	@Query(value = "SELECT  COALESCE(SUM(t_grn_gvn.grn_gvn_qty),0) FROM t_grn_gvn WHERE t_grn_gvn.fr_id=:frId AND t_grn_gvn.item_id=:itemId AND t_grn_gvn.grn_gvn_date BETWEEN :fromDate AND :toDate", nativeQuery = true)
-	int getRegTotalGrnGvn(@Param("frId") int frId, @Param("fromDate") String fromDate, @Param("toDate") String toDate,
+	float getRegTotalGrnGvn(@Param("frId") int frId, @Param("fromDate") String fromDate, @Param("toDate") String toDate,
 			@Param("itemId") int itemId);
 
 	@Query(value = "SELECT COALESCE(SUM(CASE WHEN bill_stock_type = 1 THEN qty ELSE 0 END),0) as reg , COALESCE(SUM(CASE WHEN bill_stock_type = 2 THEN qty ELSE 0 END),0) as sp FROM t_sell_bill_detail WHERE t_sell_bill_detail.item_id =:itemId AND "
@@ -27,7 +27,7 @@ public interface StockCalculationRepository extends JpaRepository<RegularSpecial
 	// grn/gvn
 		@Query(value = "SELECT COALESCE(SUM(t_grn_gvn.grn_gvn_qty),0) FROM t_grn_gvn WHERE t_grn_gvn.fr_id=:frId AND t_grn_gvn.item_id=:itemId AND"
 				+ " t_grn_gvn.grn_gvn_entry_datetime BETWEEN :fromDateTime AND :toDateTime ", nativeQuery = true)
-		int getTotalGrnGvnUptoDateTime(@Param("frId") int frId, @Param("fromDateTime") String fromDateTime, @Param("toDateTime") String toDateTime,
+		float getTotalGrnGvnUptoDateTime(@Param("frId") int frId, @Param("fromDateTime") String fromDateTime, @Param("toDateTime") String toDateTime,
 				@Param("itemId") int itemId);
 	// purchase
 	@Query(value = "SELECT  COALESCE(SUM(CASE WHEN grn_type != 3  THEN bill_qty ELSE 0 END),0) as reg ,"
